Metadata package: remove OVF-provider from default list; move retry() to Azure-OVF provider
